requirements.txt
fastapi~=0.116.1
uvicorn~=0.35.0

python-multipart~=0.0.20
psycopg2-binary~=2.9.10
SQLAlchemy~=2.0.41
pydantic-settings~=2.8.1
celery~=5.5.3
redis~=6.2.0

# pre-commit
pre-commit~=4.2.0
black~=25.1.0
isort~=6.0.1
mypy~=1.15.0

# testing
pytest~=8.4.1
pytest-celery~=1.2.0
httpx~=0.28.1
testcontainers~=4.10.0
